+++ src/sisl/physics/tests/test_spin.py
@@ -11,7 +11,7 @@ from sisl import Spin
 pytestmark = [pytest.mark.physics, pytest.mark.spin]
 
 
-def test_spin1():
+def test_spin_init():
     for val in [
         "unpolarized",
         "",
@@ -32,7 +32,7 @@ def test_spin1():
         assert s == s1
 
 
-def test_spin2():
+def test_spin_comparisons():
     s1 = Spin()
     s2 = Spin("p")
     s3 = Spin("nc")
@@ -85,95 +85,12 @@ def test_spin2():
     assert s4.is_spinorbit
 
 
-def test_spin3():
+def test_spin_unaccepted_arg():
     with pytest.raises(ValueError):
         s = Spin("satoehus")
 
 
-def test_spin4():
-    s1 = Spin(Spin.UNPOLARIZED)
-    S1 = Spin(Spin.UNPOLARIZED, np.complex64)
-    s2 = Spin(Spin.POLARIZED)
-    S2 = Spin(Spin.POLARIZED, np.complex64)
-    s3 = Spin(Spin.NONCOLINEAR)
-    S3 = Spin(Spin.NONCOLINEAR, np.complex64)
-    s4 = Spin(Spin.SPINORBIT)
-    S4 = Spin(Spin.SPINORBIT, np.complex64)
-    assert s1 == S1
-    assert s2 == S2
-    assert s3 == S3
-    assert s4 == S4
-
-    # real comparison
-    assert s1 < S2
-    assert s1 < S3
-    assert s1 < S4
-
-    assert s2 > S1
-    assert s2 < S3
-    assert s2 < S4
-
-    assert s3 > S1
-    assert s3 > S2
-    assert s3 < S4
-
-    assert s4 > S1
-    assert s4 > S2
-    assert s4 > S3
-
-    # complex complex
-    assert S1 < S2
-    assert S1 < S3
-    assert S1 < S4
-
-    assert S2 > S1
-    assert S2 < S3
-    assert S2 < S4
-
-    assert S3 > S1
-    assert S3 > S2
-    assert S3 < S4
-
-    assert S4 > S1
-    assert S4 > S2
-    assert S4 > S3
-
-    # real comparison
-    assert S1 < s2
-    assert S1 < s3
-    assert S1 < s4
-
-    assert S2 > s1
-    assert S2 < s3
-    assert S2 < s4
-
-    assert S3 > s1
-    assert S3 > s2
-    assert S3 < s4
-
-    assert S4 > s1
-    assert S4 > s2
-    assert S4 > s3
-
-    # complex complex
-    assert S1 < s2
-    assert S1 < s3
-    assert S1 < s4
-
-    assert S2 > s1
-    assert S2 < s3
-    assert S2 < s4
-
-    assert S3 > s1
-    assert S3 > s2
-    assert S3 < s4
-
-    assert S4 > s1
-    assert S4 > s2
-    assert S4 > s3
-
-
-def test_pauli():
+def test_spin_pauli():
     # just grab the default spin
     S = Spin()
 
@@ -201,7 +118,7 @@ def test_pauli():
     assert np.allclose(z, (np.conj(W) * np.dot(S.Z, W.T).T).sum(1).real)
 
 
-def test_pickle():
+def test_spin_pickle():
     import pickle as p
 
     S = Spin("nc")